requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-socketio==5.3.6
Pillow>=10.3.0
pandas==2.2.0
numpy==1.26.4
groq==0.11.0
httpx==0.27.0
python-dotenv==1.0.1
werkzeug==3.0.1
sqlalchemy==2.0.27
marshmallow==3.21.0
eventlet==0.35.2
requests==2.31.0
python-multipart==0.0.9
